[changelog][client] CDC handling newly added schema that's yet to be fetched
This edge case could happen in any of our Venice client libraries since all Venice clients
are caching the value schemas locally. Here we will adopt similar strategy as our TC that's
using RouterBackedSchemaReader to try to fetch the unknown schema from backend. System
store based metadata repository will likely be more stale than zk watcher backed cache but
luckily we are also moving to request based native metadata repository. In theory there is
still a small window where the zk cache could be stale and we still won't be able to fetch
the newly added schema that's already in use. However, in practice that should be very unlikely
since with this change and request based native metadata repo we will be on par with the
TC's RouterBackedSchemaReader guarantees and we've never run into that edge case yet.