[samza] Make ChainedPubSubCallback also take in SendMessageErrorLoggerCallback when sending messages (#1941)
During the testing of the BatchingVeniceWriter, NPE was thrown in              
VeniceWriter producer callback thread, as we always call onCompletion() API    
for the PubSubCallback.                                                        

For CompletableFutureCallback, when VW sends message, it will insert an        
internal SendMessageErrorLoggerCallback and onCompletion() will also invoke    
it. However, ChainedPubSubCallback does not register the same callback, but    
will still invoke onCompletion() API and will also invoke it's internal        
callback's onCompletion() API. If internal callback is a                      
CompletableFutureCallback, NPE will be thrown.                                 

This PR address the issue and verify through test logs that it does not        
appear and it will not throw exception if internal callback is no present.