[cc] Correctly handle concurrent polling (#1945)
Problem Statement
The CDC client is facing this exception:
java.util.ConcurrentModificationException: KafkaConsumer is not safe for multi-threaded access

Solution
This is because two threads are invoking poll, but we don't handle it correctly. Inside internalPoll, we first call subscriptionLock.writeLock().tryLock(timeoutInMs, TimeUnit.MILLISECONDS). However, if we don't acquire the lock within timeoutInMs, we still call pubSubConsumer.poll(timeoutInMs). This leads to the exception in the KafkaConsumer. To fix this, we check if we acquired the lock. If we didn't then we return immediately with an empty collection.