[common] Relex validation checks when annotating value schema
Today, when annotating a value schema, we use the STRICT requirement to validate the input
value schema, see class SchemaParseConfiguration for the details of what these validations are.
However, we have seen for some stores that its value schemas that are not in strictly valid Avro
schemas, but they are already in production use. For example, some might have a double type field
whose default value set to "0" instead of "0.0", or unions whose default value does not match the
1st branch in the union. The schema might not be easily removed because for hybrid stores, it's
possible that some producer still use it to produce.

We change the function in SchemaUtils to use loose validation to allow the schema to be parsed even
if it has some minor issues.