[controller] clean up old store version in parent controller (#1958)
When venice repairs a system store in parent controller, it calls addVersionAndTopicOnly but it does not clean up the old. 
store versions. Due to repeated store repair this can lead to explosion in the zk node of the store config which can choke. 
any operations with the store. This PR cleans up old version while repairing system stores.