[common] Relax validation checks when annotating value schema (#1960)
* [common] Relax validation checks when annotating value schema 

Today, when annotating a value schema, we use the STRICT requirement to validate the input
value schema; see class SchemaParseConfiguration for the details of these validations.
However, we have seen for some stores that their value schemas are not strictly valid Avro
schemas, but are already in production use. For example, some might have a double type field
whose default value is set to "0" instead of "0.0", or unions whose default value does not match the
1st branch in the union. The schema might not be easily removed because, for hybrid stores, it's
possible that some producers still use it to produce.

We added a new function api in SchemaUtils to provide an option to use loose validation to allow the schema to be parsed even if it has some minor issues.