[fast-client] delay version switch until the newly fetch version is ready (#1970)
* [fast-client] delay version switch until the newly fetch version is ready

Today, in Fast Client, during the metadata update, once a new version
is discovered, Fast Client will switch to the new current version
right away even though the routing info for the new current version
doesn't contain replicas for every partition and this can happen
as the metadata fetched is from Venice Server, which can be suffering
from a delay from Zookeeper since Controller is making a decision
about when to switch to the new version.

This PR will delay the current version switch until the returned
routing info contains at least some ready-to-serve replicas for
every partition.
There is an edge case when the being-used current version gets
deleted, even the routing info is not complete, Fast Client
will switch to the new version.

TODO: improve Venice Server metadata endpoint to force a refresh
when the routing info is incomplete.