[server] Initialize follower HB entry before waiting for latch to complete (#1986)
When using heartbeat lag for ready-to-serve check, we identified this issue: HB follower entry was not initialized in the participant model when going from offline to standby and current version replica bootstrap times out waiting for latch to complete. This is a dead lock in behavior and we should adjust the order.