[Controller] Initialize venice_parent_metadata_system_store for standby controllers (#1938)
As part of enabling high availability for the Parent Controller, this commit sets up an in-memory system store to hold metadata for all Venice clusters. This store will be accessible to future standby parent controllers in other colos, laying the foundation for failover support and preventing race conditions during controller transitions.