[build] Make "all-modules" be published after all other modules are published (#1995)
We have a special gradle module called "all-modules". This is supposed to be a single-module to add watchers on to know when new versions of Venice are released. By default, Gradle publish plugin publishes to Maven repo in alphabetical order. Due to it's name, "all-modules" is the first module that gets published.

By ensuring that "all-modules" gets published last, we can use the same watcher to also ensure that all other modules have been published successfully, and the published version is ready to be used.