[controller] Add execution id in the header of admin topic messages (#1988)
In future, we want to rely on execution id to skip admin messages. For this purpose, I am adding execution id in the
header of the kafka messages that are meant to be written into admin topic. In future, we will use execution id from this
header, instead of from the payload, and use it for various purpose, including skip message.
This will also help in the case, where the payload cannot be deserialized.
- also did some refactoring to cache LeaderCompleteState headers; this will reduce GC happening because of creation
of new LeaderCompleteState headers for every single message.