[controller] Support configuring automatic deregistration of Controller Helix participant (#2013)
When we run controllers as Deployments on Kubernetes, the hostname is different everytime the pod moves. Since Venice controllers register themselves as participants of a Helix cluster, the list of "offline" instances grows indefinitely till it is cleaned up. Helix supports a functionality to automatically deregister participants of a cluster if they have been offline for a configured period of time. LinkedIn has been operating with this config in production for a few months, but it needs to be configured as an operational task after the cluster has been created. This change allows this to be configured at time of controller cluster creation.