[controller] Remove reflection from TestZkHelixAdminClient (#2016)
"TestZkHelixAdminClient" uses reflection to mock and validate certain configs. However, this is due to the fact that it relies on "VeniceControllerClusterConfig" instead of "VeniceControllerMultiClusterConfig". This change adds additional methods in "VeniceControllerMultiClusterConfig" and that helps make mocking easier in tests in a way where we can completely remove use of reflection.