[controller] Support configuring cluster topology when creating storage clusters in Helix (#2015)
Helix supports topology-aware partition placement, and it is being widely used in production at LinkedIn. However, configuring this is an operational task, and done after the cluster is created. This change allows configuring the topology-aware configs of the Helix cluster at time of cluster creation, so that it can be handled automatically.

Also cleaned up some code and tests that previously configured this incorrectly