[cc] Add consumer sequence id to VCC for order comparison (#2012)
* [cc] Add consumer sequence id to VCC for order comparison

Problem: records consumed from a change log consumer may undergo some sort of
parallel processing/transformation. When collecting/merging the results of these
parallel processing we may encounter records for the same key with the same
timestmap. These records were ordered in the change log but that order is shuffled
or lost during the parallel processing.

Solution: provide a monotonically increasing consumer sequence id to VeniceChangeCoordinate
(VCC) which is also serializable and provided for every record consumed from the change log
consumer. The sequence id from the same consumer instance and partition can be compared to
determine the ordering and "distance" of the records/events.

The starting sequence id is seeded by the starting timestamp of the consumer in ns.
In theory it's possible for sequence id to overlap when a restart occurs which may
require the stream processing pipeline to be fully drained between restarts. However,
in practice this is very unlikely even with time drift. It would require restart to take
less than 1ms and a consumption rate of 10^9 records per second per partition.