[router] Reduce the recording frequency for pending_request_count per route (#2031)
The pending_request_count per route is quite expensive since
it would be recorded for every storage node request.
Previously, the metric was recorded for every host health check
in {@link VeniceHostHealth#isHostHealthy}, which is not
necessary, this logic would amplify the recording frequency
by the number of replicas as {@link VeniceHostFinder} will
invoke this function for every replica during host selection
for a given partition and this will become especially expensive
in high throughput cluster.
The new logic will only record this metric when sending out
the storage node request and it doesn't record again when
the request is completed to reduce the frequency.
We are mainly interested in how high the pending request count
can be, so a minor inaccuracy won't matter.