[controller] System store repair service should not fix store that retrieves correct heartbeat after retries (#2034)
Today system store repair service will periodically fetch the heartbeat from   
store it is checking. The problem is if initial heartbeat retrieved is stale   
or unavailable, it would not remove it from the unhealthy store set that it    
plans to fix. This leads to unnecessary push job to the store.                  

Also, remove retry in repair process as it is not necessary and should be      
taken care of by periodic check.