[controller] Backup version storage optimization (#2029)
Venice maintains a backup version for every store, which is only used in rare cases when a user explicitly rolls back to it.  
Since such rollbacks are rare, we can optimize resource usage by lowering the replication factor of these backup  
versions.
This change introduces a mechanism to reduce the replication factor of the backup version one hour after a new current
 version is created. Doing so helps minimize unnecessary storage overhead, while also reducing the risk of large-scale  
 rebalancing immediately after a push.
Incase if the reduction caused performance degradation, we can use helix endpoint to increase the replication factor   back to original RF number.

https://helix.apache.org/1.4.2-docs/tutorial_rest_service.html