[vpj] Improve logging in VPJ for largest uncompressed record size (#2052)
Today, when VPJ logs the larges uncompressed value size, it logs in bytes. However, it is beneficial to log this in human-readable form for better UX.

This commit logs the largest uncompressed value bytes size in human-readable form. To help with internal debugging, also logged the raw bytes value in the Spark accumulator section