[da-vinci] Added key urn compression support in DaVinci
Today, quite a number of customers within LinkedIn are using URNs as key
schema and there is an issue with URNs since they contain common prefix:
`urn:li:` + urn type with a following unique id. The total number of urn
types are limited.
In some use cases, the value is relatively small comparing to the key schema
since they only keep a few of scalar types and this is typically OK if
the stores are being served out of remote Venice within disk-backed clusters,
but it will consume a lot of memory resources in DaVinci or memory-serving
clusters.
In DaVinci, by default, we will keep all the data in RAM and it will use
PlainTable format.
RocksDB PlainTable format doesn't support compression as it is optimized
for fast lookup.
To reduce the memory usage, this PR introduces auto URN compression
in key schema since value compression can be done via today's Venice compression.
The main reason we don't want to leverage the general compression lib
is that we need to keep the compressed bytes consistent when evolving
the compression algo and there is no such guarantee with the public
compression libs.

There are two layers of control:
1. Store level and store owners can enable this feature and specify
   the urn fields (only top-level).
2. DaVinci application-level.
The store will enable the new feature when both are true.

So far, this feature is not compatible with the following features:
1. Blob transfer.
2. Partial update in DaVinci.

We will have a follow-up PR to support #1 and for #2, it might not
be necessary.