[controller] Improve davinci debugging by adding host information. (#1977)
Currently, when a Venice Push Job (VPJ) hangs for a specific region, it's often due to one or a
small subset of partitions not completing. However, VPJ does not explicitly surface which host
with which partition is causing the delay. As a result, on-call engineers must investigate
multiple sources—including controller and server logs, ZooKeeper—to identify the blocking
partition and server.

One example logging:
1827/1840 Da Vinci instances completed.. However, some instances are still reporting partition
level status keys and they are not completed yet. 19/20 partitions completed in 7318 Da Vinci
replicas.. Following partitions still not complete (capped at 10) [0]

This challenge is even more pronounced for Da Vinci, where identifying the problematic host and
partition requires additional human toil to query the system store with specific cmd.