[cc] Change Version Swap Count to UpDownCounter in OTEL and Total in Tehuti (#2091)
Problem Statement
The OTEL metric VERSION_SWAP_COUNT, wasn't being emitted every minute.

Solution
To fix the emission for VERSION_SWAP_COUNT, it will use a UP_DOWN_COUNTER in OTEL and Total in Tehuti. For UP_DOWN_COUNTER, it will only start emitting once a value is recorded to it. To fix this, we record a 0 in the constuctor of BasicConsumerStats.

We cannot use a Gauge for OTEL, because a synchronous gauge doesn't automatically emit the last recorded value. Only the asynchronous gauge does, but getting it to work with multiple dimensions isn't pretty. We can technically have the background reporter thread in CDC emit the Gauge value, but that is hacky. UP_DOWN_COUNTER provided the cleanest solution, so we will be going with that.