[controller] Skip admin message by execution id (#2060)
We are replacing `skipping admin messages by offset` with `skipping admin messages by execution id` because offset is not available in all pubsub systems.
During the rollout period, skipping admin messages will be supported by both offset and execution id (but not both in the same admin tool command) and hence there is some duplicate code, e.g. `checkOffsetToSkip()` and `checkExecutionIdToSkip()`, `skipMessageWithOffset()` and `skipMessageWithExecutionId()`.
This is to allow users to transition from using offset to using execution id.
In near future, we will remove the support for skipping by offset.