[server] Remove unused replica progress reporting in ReplicaStatus (#2100)
Progress was previously reported as a numeric offset. While this was easy to  
interpret, it was rarely, if ever, used in practice. With the new PubSubPosition,  
progress is represented as raw bytes, which are not meaningful or helpful for  
troubleshooting. Instead of adding extra code to report this unused information,  
the reporting logic is removed.