[server][da-vinci] Check for closed writer before send (#2085)
During KILL message processing, Venice closes the writers, while there could be messages in the producer buffer.   
This might cause exceptions such as The internal KafkaProducer has been closed: This causes the exception in the  
 store ingestion task which is later treated as ingestiion error which is misleading.
This PR checks for close writer during processing of messages and also checks in a SIT is running before reporting error.