[vpj] Fix AbstractPartitionWriter to pass the correct rmdSchemaId (#2103)
As part of support for logical timestamp and RMD support for VPJ, we did not pass the right rmd schema id to the Writer and as a result the payload written to the server did not carry any writer schema information.
As a result, any subsequent consumption of the RMD data would result in NPE due to lack of writer schema id.

Pass the rmd schema id as part of processing input records prior to writing to Kafka
Add a check to AbstractPartitionWriterMessage to ensure version id cannot be < 0 in case of non-null RMD payloads