[dvc][doc] Improve documentation and usability for DVRT (#2106)
Problem Statement
The documentation for DVRT is stale and there's some feature gaps.

Solution
This PR overhauls the JavaDocs and the website documentation (https://kvargha.com/venice/docs/user_guide/read_api/da_vinci_client#record-transformer-dvrt) for DVRT. This should make it easier for newcomers to onboard.

To improve usability, we include the storeName in the DVRT constructor to make general implementations be applicable to multiple stores dynamically.