[fc][dvc][controller][producer] Match latest canonical value schema if there's no exact match (#2117)
When there are multiple value schemas having the same canonical schema, in many places, the data is written with any schema ID
that has the same canonical schema. This was okay in the past because clients would return data in the latest value schema
registered on the store. Recently, we changed the behavior to return the data in the same schema that the user wrote the data
in, and since the schema id didn't match the exact schema that the user-provided, this would cause issues in the read path for
some cases.

This commit changes the behavior to first try to match the exact schema, and only fall back to the canonical schema if an exact
schema match cannot be found. If there are multiple schemas that map to the same canonical schema, then the latest value schema
id will be used.

This behavior was already present in some parts of the system. This change unifies all implementations to use the same helper
function and hence, the behavior should now be consistent across components.