[venice-thin-client] Cache value schema to avoid repeated calls to generate canonical form (#2112)
In Venice client today we convert a value schema to its canonical form on every request, which causes excessive  
memory usage in the client app. This PR caches the record deserializer in the client level, so that any any given  
schema id it will generate the canonical schema at most once.