[server] Add InactiveTopicPartitionChecker to pause/resume inactive topic partition periodically (#2152)
We observed that when two topic-partitions with the same leader Kafka broker   
are assigned to the same consumer, if one is idle, it may impact fetch         
performance of the other topic-partition.                                      

To address this issue, a few ideas have been proposed offline.                  
This PR implements one of the solution: Introduce a checker service that       
periodically check current version pool's consumers and pause idle             
topic-partition for some times.                                                

For now, the check works as follows:                                           
The check will be scheduled every                                             
server.inactive.topic.partition.checker.internal.in.seconds seconds.           
In the check, it will examine every topic-partition that, if delta between     
previous message and now is beyond certain threshold                          
"server.inactive.topic.partition.checker.threshold.in.seconds", it will try    
to add them to the pause set. It will be resumed by the next check cycle and   
in next check cycle it won't be paused (avoid infinite paused).