[server] Fix concurrent modification in map iteration (#2186)
This PR fixes a simple concurrent modification in InactiveTopicChecker. We cannot iterate the set while removing the entry.
This PR changes it to iterate item first, then clear the set.
Also, make some logging changes so one log covers one action of a consumer task and make then INFO level, as this is expected behavior of the class.