[protocol] introduce new ETL strategy config (#2183)
Adding a new etl strategy config to control the ETL behavior of Venice. Currently
the ETL behavior is completely external. Meaning it's a completely separate service
and external to Venice. Meaning it will detect and manage the Venice store versions
via controller client APIs or other means. This could change in the future and the
new config will be used to indicate the ETL strategy.