[server][dvc] Initiate no-op KeyUrnCompressionFunction for storage engine (#2206)
We found error requests in during server restart case.
This is because during server restart, storage engine got restored first and listener service restored before ingestion is properly restarted. Due to the fact that key urn compression function is created during SIT construction, so that it will be initiated to null for storage engine in the beginning. This results in null check error in storage engine read operation when it is looking for key urn compression function before SIT construction.

The tactical fix here is to initiate it to NoOp instead of null to bypass the check.
(1) Server: It will handle request properly.
(2) Da Vinci without liveUpdateSuppression: It will create SIT and ingest to ready-to-serve before serving read.
(3) Da Vinci with liveUpdateSuppression: Do not enable key urn compression feature. With this change, it will also handle the read properly. Need proper fix for feature gap.