[server] Fix entry retrieval bug in previously ready-to-serve state checkpoint deserialization (#2214)
In Java, the default Avro string type deserialization returns a Utf8 type, instead of String type, unless specified in Avro compilation parameter.

This PR fixes the bug that the previously stored replica ready-to-serve state in OffsetRecord(PartitionState) is not successfully retrieved. It tries to add a Java String-String key-value pair and try to retrieve it during restart to decide whether fast-restart is allowed. However, the deserialized map is actually a Utf8-Utf8, so the key lookup will fail and the feature is never enabled for offset-based fast restart or (heartbeat) time-base fast restart.