[server][dvc] Increase  checksum validation threadpool worker thread count for blob receiver (#2218)
Fix incorrect async checksum validation threadpool setup in blob transfer      

This fixes the incorrect setup in #2197 for the async checksum validation      
threadpool for the blob transfer receiver. The original setup is wrong         
because we introduced an unbounded queue to make sure all the requests are     
started, but then since core size is 0, it will not create more threads until  
the queue size is full. This results in only 1 thread handling the requests    
sequentially and thus slowing down everything.                                 

The fix is simple: just default the number of available cores should be good   
enough for now. If we find during future testing that we need to adjust this,  
we can add more config to tune it.                                             

Performed a few tests — this runs as fast as #2213 in the same cluster setup   
with enhanced safeguards. Average validation time is 100~200 ms for SST files  
of size 32–256 MB. Total transfer time is ~400 s for 8 GB with 50–60 parallel  
transfers, which is good enough.                                               

A future improvement is to migrate MD5 checksum to CRC32 checksum, but that    
will not happen in this PR.