[server] Move current version rebalance replica into non-current pool (#2285)
In today's setup, current version will always go to current version pools no    
matter it is ready-to-serve or not. This introduces disruptive ingestion       
volume in the current version consumer pool, as high consumption rate might    
slow down or starve current version ingestion. The true north is using         
replica checkpoint (blob transfer) to replace the pub-sub re-ingestion         
traffic, but before it is ready and ramped, we need a minimal effort solution  
to offset the disruption.                                                      

This PR aims to solve it by moving the current version not-ready-to-serve      
replica into the non-current version pool. It will track whether a partition   
has resubscribed yet in PCS, and when current version replica becomes          
completed, we will invoke a resubscribe() to move it to the current version    
pool.                                                                          

Also, the current version max lag signal used by the adaptive throttling       
system will be adjusted accordingly. It will only compute ready-to-serve       
current version replicas.                                                      

This PR also introduces a new config to control the check frequency, as it is  
not needed to check on every SIT thread loop.                                  

server.resubscription.check.interval.in.seconds controls the interval and      
default interval is 10s.