[dvc] Fix Da Vinci heartbeat recording region name (#2294)
Fix inconsistent region recording causing Da Vinci not ready-to-serve          

During test environment testing, we found that Da Vinci is not able to come    
to ready-to-serve when heartbeat-based ready-to-serve check is enabled.        
Through the heapdump we found that the recording logic is inconsistent with    
the checking logic:                                                            
1. Recording hardcoded "" as region for DVC                                       
2. Checking logic refers to serverConfig.regionName for follower max lag.         

This PR fixes the recording logic so that it uses the same config for          
recording, ensuring consistency with the checking logic.