[server][dvc] Catch offset lag measurement exception in ingestion path (#2313)
There are still lingering usages of offset-lag measurement in the ingestion     
path:                                                                           
(1) Fast restart (will be disabled in this PR)                                 
(2) Blob Transfer (will be disabled with time-lag logic)                       

Thus, we are still checkpointing offset lag in the sync offset logic. When     
the dependent PubSub system explodes and throws an exception in the offset-    
lag measurement code, it throws an exception in the drainer and causes all     
related replicas to go into error state. Even though we have some reset        
logics, they do not cover all cases. We should avoid such errors in the first  
place to reduce operational burden.                                            

This PR adds:                                                                   
(1) Catching exceptions in hybrid offset-lag measurement logic and using a     
    redundant filter to control logging volume.                                
(2) A new config to disable offset-lag checkpointing in the sync offset        
    logic:                                                                      
    offset.lag.checkpoint.during.sync.enabled, default true to keep current    
    behavior. This will be manually disabled during rollout.                   
(3) Adjusted fast restart behavior: when time-lag based ready-to-serve is      
    enabled and time-lag based fast restart is not enabled, we should not      
    fall back to offset-lag checkpointing, as it mixes feature behaviors.