[server] Add a heartbeat lag based replica auto resubscribe feature (#2269)
This is an optional feature that aims to improve consumption resiliency. When  
we observe current-version resource heartbeat delay growing beyond a certain    
threshold, it may indicate that the underlying consumer is not functioning     
well. While we are also working on identifying the root cause, this feature    
aims to mitigate the issue at runtime by asking SIT to attempt a resubscribe   
action that moves the replica to another consumer in the same consumer pool.