[vpj] [controller] Fix categorization of various errors in VPJ (#2308)
The solution is to handle these errors appropriately in VenicePushJob and invoking the correct PushJobCheckpoint for these errors.

Note: We can follow up with PR to move the responsibility of checkpointing to the exceptions themselves. It consolidates the categorization and avoids scattered error handling which is prone to missing scenarios and edge cases.

Code changes
Change controller createVersion code path to return new errorType for ACL issues
Introduce new errorType ACL_ERROR
Introduce new exceptions for store ACL failures and for shema mismatchs
---------

Co-authored-by: Bharath Kumarasubramanian <bkumaras@bkumaras-mn2.linkedin.biz>