[server][common] Make server lag monitor cleanup cycle configurable (#2316) (#2324)
Problem: In practice we are seeing that the local customized view cache can be out of sync or stale for an extended period of time for large clusters or when Helix performance is degraded. As a result the hard coded default of 5 for DEFAULT_LAG_MONITOR_CLEANUP_CYCLE which translates to approximately 5 minutes (5 * DEFAULT_LAG_LOGGING_THREAD_SLEEP_INTERVAL_SECONDS = 60) is not sufficient and we are seeing false positive cleanups. e.g. an intended assignment's heartbeat monitor was removed by mistake due to CV is stale for more than 5 minutes. As a result the lag based replica readiness check can never complete.

Solution: Make the lag monitor cleanup cycle configurable so we can tune it accordingly while tackling performance issues to reduce the out of sync window for customized view.

Co-authored-by: Xun Yin <xunyin8@gmail.com>